+++ main.cpp
@@ -32,6 +32,7 @@
 #include "MatchList.h"
 #include "FileScanner.h"
 #include "OutputTask.h"
+#include "MemDiags.h"
 
 // This is for gnulib's error.c.
 char *program_name;
@@ -112,6 +113,12 @@ int main(int argc, char **argv)
 		// Wait for the output thread to complete.
 		ot.join();
 
+#if MEMUSE_DEBUG
+		// Dump memory stats.
+		MemDiags md;
+		md.PrintStats();
+#endif // MEMUSE_DEBUG
+
 		// Check for errors.
 		if(g.Error())
 		{
@@ -144,5 +151,18 @@ int main(int argc, char **argv)
 	{
 		std::cerr << "ucg: Error during arg parsing: " << e.what() << std::endl;
 		return 255;
+	}
+	catch(const std::runtime_error &e)
+	{
+		std::cerr << "ucg: std::runtime_error exception: " << e.what() << std::endl;
+		return 255;
+	}
+	catch(...)
+	{
+		// We shouldn't need this catch(...), but Cygwin seems to not properly call the default
+		// terminate handler (which should be abort()) if we let an exception escape main(), but will simply return
+		// without an error code.  I ran into this when trying to instantiate std::locale with locale=="" in ArgParse,
+		// and before I had the std::runtime_error catch clause above.
+		abort();
 	}
 }