+++ src/File.h
@@ -1,18 +1,18 @@
 /*
  * Copyright 2015 Gary R. Van Sickle (grvs@users.sourceforge.net).
  *
- * This file is part of UltimateCodeSearch.
+ * This file is part of UniversalCodeGrep.
  *
- * UltimateCodeSearch is free software: you can redistribute it and/or modify it under the
+ * UniversalCodeGrep is free software: you can redistribute it and/or modify it under the
  * terms of version 3 of the GNU General Public License as published by the Free
  * Software Foundation.
  *
- * UltimateCodeSearch is distributed in the hope that it will be useful, but WITHOUT ANY
+ * UniversalCodeGrep is distributed in the hope that it will be useful, but WITHOUT ANY
  * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
  * PARTICULAR PURPOSE.  See the GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License along with
- * UltimateCodeSearch.  If not, see <http://www.gnu.org/licenses/>.
+ * UniversalCodeGrep.  If not, see <http://www.gnu.org/licenses/>.
  */
 
 /** @file */
@@ -46,6 +46,12 @@ public:
 
 	const char * data() const { return m_file_data; };
 
+	/**
+	 * Returns the name of this File as passed to the constructor.
+	 * @return  The name of this File as passed to the constructor.
+	 */
+	std::string name() const { return m_filename; };
+
 private:
 
 	/**
@@ -67,6 +73,8 @@ private:
 	 * @param file_size  The same file_size used in the call to GetFile().
 	 */
 	void FreeFileData(const char * file_data, size_t file_size);
+
+	std::string m_filename;
 
 	int m_file_descriptor { -1 };