+++ src/FileScanner.cpp
@@ -134,7 +134,7 @@ void FileScanner::Run()
 
 	// Pull new filenames off the input queue until it's closed.
 	std::string next_string;
-	while(m_in_queue.wait_pull(next_string) != queue_op_status::closed)
+	while(m_in_queue.wait_pull(std::move(next_string)) != queue_op_status::closed)
 	{
 		MatchList ml(next_string);
 
@@ -161,8 +161,8 @@ void FileScanner::Run()
 
 			if(!ml.empty())
 			{
-				/// @todo Move semantics here?
-				m_output_queue.wait_push(ml);
+				// Force move semantics here.
+				m_output_queue.wait_push(std::move(ml));
 			}
 		}
 		catch(const FileException &error)
@@ -365,6 +365,6 @@ void FileScanner::ScanFileLibPCRE(const char *file_data, size_t file_size, Match
 		prev_lineno = line_no;
 		Match m(file_data, file_size, ovector[0], ovector[1], line_no);
 
-		ml.AddMatch(m);
+		ml.AddMatch(std::move(m));
 	}
 }