+++ src/sync_queue.h
@@ -1,18 +1,18 @@
 /*
- * Copyright 2015 Gary R. Van Sickle (grvs@users.sourceforge.net).
+ * Copyright 2015-2016 Gary R. Van Sickle (grvs@users.sourceforge.net).
  *
- * This file is part of UltimateCodeSearch.
+ * This file is part of UniversalCodeGrep.
  *
- * UltimateCodeSearch is free software: you can redistribute it and/or modify it under the
+ * UniversalCodeGrep is free software: you can redistribute it and/or modify it under the
  * terms of version 3 of the GNU General Public License as published by the Free
  * Software Foundation.
  *
- * UltimateCodeSearch is distributed in the hope that it will be useful, but WITHOUT ANY
+ * UniversalCodeGrep is distributed in the hope that it will be useful, but WITHOUT ANY
  * WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
  * PARTICULAR PURPOSE.  See the GNU General Public License for more details.
  *
  * You should have received a copy of the GNU General Public License along with
- * UltimateCodeSearch.  If not, see <http://www.gnu.org/licenses/>.
+ * UniversalCodeGrep.  If not, see <http://www.gnu.org/licenses/>.
  */
 
 /** @file */
@@ -79,7 +79,7 @@ public:
 		return queue_op_status::success;
 	}
 
-	queue_op_status wait_push(const ValueType&& x)
+	queue_op_status wait_push(ValueType&& x)
 	{
 		std::lock_guard<std::mutex> lock(m_mutex);
 
@@ -122,7 +122,38 @@ public:
 		}
 
 		// Otherwise, we have something in the queue to pull off.
+		x = m_underlying_queue.front();
+		m_underlying_queue.pop();
+
+		return queue_op_status::success;
+	}
+
+	queue_op_status wait_pull(ValueType&& x)
+	{
+		// Using a unique_lock<> here vs. a lock_guard<> because we'll be using a condition variable, which needs
+		// to unlock the mutex.
+		std::unique_lock<std::mutex> lock(m_mutex);
+
+		if(m_underlying_queue.empty() && m_closed)
+		{
+			// Queue is empty and has been closed, let the caller know.
+			return queue_op_status::closed;
+		}
+
+		// Wait until the queue is not empty, or somebody closes the sync_queue<>.
+		m_cv.wait(lock, [this](){ return !m_underlying_queue.empty() || m_closed; });
+
+		// Check if we've be awoken to a closed and empty queue.
+		if(m_underlying_queue.empty() && m_closed)
+		{
+			// We have, let the caller know.
+			return queue_op_status::closed;
+		}
+
+		// Otherwise, we have something in the queue to pull off.
 		// Use move-assignment vs. copy-assignment for efficiency.
+		// Note that C++11 std::queue<>::front() returns a non-const reference as well as a const one, so
+		// std::move() will work here.
 		x = std::move(m_underlying_queue.front());
 		m_underlying_queue.pop();